+++ diffractio/scalar_fields_XYZ.py
@@ -598,14 +598,14 @@ class Scalar_field_XYZ(object):
         deltaz = self.z[1] - self.z[0]
         rangox = self.x[-1] - self.x[0]
         rangoy = self.y[-1] - self.y[0]
-        pixelx = np.linspace(-numx / 2, numx / 2, numx)
+        pixelx = np.linspace(-int(numx / 2), int(numx / 2), numx)
         pixely = np.linspace(-numy / 2, numy / 2, numy)
 
         modo = self.u0.u
 
         # Calculo de la phase 1 normalizada -------------------
-        kx1 = np.linspace(0, numx / 2 + 1, int(numx / 2))
-        kx2 = np.linspace(-numx / 2, -1, int(numx / 2))
+        kx1 = np.linspace(0, int(numx / 2) + 1, int(numx / 2))
+        kx2 = np.linspace(-int(numx / 2), -1, int(numx / 2))
         kx = (2 * np.pi / rangox) * np.concatenate((kx1, kx2))
 
         ky1 = np.linspace(0, numy / 2 + 1, int(numy / 2))
